requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
abseil-cpp=20211102.0=hd4dd3e8_0
absl-py=1.3.0=py38h06a4308_0
aiohttp=3.8.3=py38h5eee18b_0
aiosignal=1.2.0=pyhd3eb1b0_0
appdirs=1.4.4=pyhd3eb1b0_0
asttokens=2.2.1=pyhd8ed1ab_0
async-timeout=4.0.2=py38h06a4308_0
attrs=22.1.0=py38h06a4308_0
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=pyhd8ed1ab_3
backports.functools_lru_cache=1.6.4=pyhd8ed1ab_0
blas=1.0=mkl
blinker=1.4=py38h06a4308_0
bottleneck=1.3.5=py38h7deecbd_0
brotli=1.0.9=h5eee18b_7
brotli-bin=1.0.9=h5eee18b_7
brotlipy=0.7.0=py38h27cfd23_1003
bzip2=1.0.8=h7b6447c_0
c-ares=1.19.0=h5eee18b_0
ca-certificates=2023.05.30=h06a4308_0
cachetools=4.2.2=pyhd3eb1b0_0
certifi=2023.7.22=py38h06a4308_0
cffi=1.15.1=py38h5eee18b_3
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.0.4=py38h06a4308_0
contourpy=1.0.5=py38hdb19cb5_0
cryptography=39.0.1=py38h9ce1e76_0
cuda-cudart=11.7.99=0
cuda-cupti=11.7.101=0
cuda-libraries=11.7.1=0
cuda-nvrtc=11.7.99=0
cuda-nvtx=11.7.91=0
cuda-runtime=11.7.1=0
cycler=0.11.0=pyhd3eb1b0_0
dbus=1.13.18=hb2f20db_0
debugpy=1.5.1=py38h295c915_0
decorator=5.1.1=pyhd8ed1ab_0
entrypoints=0.4=pyhd8ed1ab_0
executing=1.2.0=pyhd8ed1ab_0
expat=2.4.9=h6a678d5_0
ffmpeg=4.3=hf484d3e_0
filelock=3.9.0=py38h06a4308_0
fontconfig=2.14.1=h4c34cd2_2
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.12.1=h4a9f257_0
frozenlist=1.3.3=py38h5eee18b_0
giflib=5.2.1=h5eee18b_3
glib=2.69.1=he621ea3_2
gmp=6.2.1=h295c915_3
gmpy2=2.1.2=py38heeb90bb_0
gnutls=3.6.15=he1e5248_0
google-auth=2.6.0=pyhd3eb1b0_0
google-auth-oauthlib=0.4.4=pyhd3eb1b0_0
grpc-cpp=1.48.2=h5bf31a4_0
grpcio=1.48.2=py38h5bf31a4_0
gst-plugins-base=1.14.1=h6a678d5_1
gstreamer=1.14.1=h5eee18b_1
icu=58.2=he6710b0_3
idna=3.4=py38h06a4308_0
importlib-metadata=6.0.0=py38h06a4308_0
importlib_resources=5.2.0=pyhd3eb1b0_1
intel-openmp=2021.4.0=h06a4308_3561
ipykernel=6.14.0=py38h7f3c49e_0
ipython=8.4.0=py38h578d9bd_0
jedi=0.18.2=pyhd8ed1ab_0
jinja2=3.1.2=py38h06a4308_0
joblib=1.1.1=py38h06a4308_0
jpeg=9e=h5eee18b_1
jupyter_client=7.0.6=pyhd8ed1ab_0
jupyter_core=4.12.0=py38h578d9bd_0
kiwisolver=1.4.4=py38h6a678d5_0
krb5=1.19.4=h568e23c_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libbrotlicommon=1.0.9=h5eee18b_7
libbrotlidec=1.0.9=h5eee18b_7
libbrotlienc=1.0.9=h5eee18b_7
libclang=14.0.6=default_hc6dbbc7_1
libclang13=14.0.6=default_he11475f_1
libcublas=11.10.3.66=0
libcufft=10.7.2.124=h4fbf590_0
libcufile=1.6.1.9=0
libcurand=10.3.2.106=0
libcusolver=11.4.0.1=0
libcusparse=11.7.4.91=0
libdeflate=1.17=h5eee18b_0
libedit=3.1.20221030=h5eee18b_0
libevent=2.1.12=h8f2d780_0
libffi=3.4.2=h6a678d5_6
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=11.2.0=h00389a5_1
libgfortran5=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libiconv=1.16=h7f8727e_2
libidn2=2.3.2=h7f8727e_0
libllvm14=14.0.6=hdb19cb5_3
libnpp=11.7.4.75=0
libnvjpeg=11.8.0.2=0
libpng=1.6.39=h5eee18b_0
libpq=12.9=h16c4e8d_3
libprotobuf=3.20.3=he621ea3_0
libsodium=1.0.18=h36c2ea0_1
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.19.0=h5eee18b_0
libtiff=4.5.0=h6a678d5_2
libunistring=0.9.10=h27cfd23_0
libuuid=1.41.5=h5eee18b_0
libwebp=1.2.4=h11a3e52_1
libwebp-base=1.2.4=h5eee18b_1
libxcb=1.15=h7f8727e_0
libxkbcommon=1.0.1=h5eee18b_1
libxml2=2.10.3=hcbfbd50_0
libxslt=1.1.37=h2085143_0
lz4-c=1.9.4=h6a678d5_0
markdown=3.4.1=py38h06a4308_0
markupsafe=2.1.1=py38h7f8727e_0
matplotlib=3.7.1=py38h06a4308_1
matplotlib-base=3.7.1=py38h417a72b_1
matplotlib-inline=0.1.6=pyhd8ed1ab_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py38h7f8727e_0
mkl_fft=1.3.1=py38hd3c417c_0
mkl_random=1.2.2=py38h51133e4_0
mpc=1.1.0=h10f8cd9_1
mpfr=4.0.2=hb69a4c5_1
mpmath=1.2.1=py38h06a4308_0
multidict=6.0.2=py38h5eee18b_0
munkres=1.1.4=py_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.5.6=pyhd8ed1ab_0
nettle=3.7.3=hbbd107a_1
networkx=2.8.4=py38h06a4308_1
nspr=4.33=h295c915_0
nss=3.74=h0370c37_0
numexpr=2.8.4=py38he184ba9_0
numpy=1.24.3=py38h14f4228_0
numpy-base=1.24.3=py38h31eccc5_0
oauthlib=3.2.2=py38h06a4308_0
openh264=2.1.1=h4ff587b_0
openssl=1.1.1u=h7f8727e_0
packaging=23.0=py38h06a4308_0
pandas=1.5.3=py38h417a72b_0
parso=0.8.3=pyhd8ed1ab_0
pcre=8.45=h295c915_0
pexpect=4.8.0=pyh1a96a4e_2
pickleshare=0.7.5=py_1003
pillow=9.4.0=py38h6a678d5_0
pip=23.0.1=py38h06a4308_0
ply=3.11=py38_0
pooch=1.4.0=pyhd3eb1b0_0
prompt-toolkit=3.0.38=pyha770c72_0
protobuf=3.20.3=py38h6a678d5_0
psutil=5.9.0=py38h5eee18b_0
ptyprocess=0.7.0=pyhd3deb0d_0
pure_eval=0.2.2=pyhd8ed1ab_0
pyasn1=0.4.8=pyhd3eb1b0_0
pyasn1-modules=0.2.8=py_0
pycparser=2.21=pyhd3eb1b0_0
pygments=2.15.1=pyhd8ed1ab_0
pyjwt=2.4.0=py38h06a4308_0
pyopenssl=23.0.0=py38h06a4308_0
pyparsing=3.0.9=py38h06a4308_0
pyqt=5.15.7=py38h6a678d5_1
pyqt5-sip=12.11.0=py38h6a678d5_1
pysocks=1.7.1=py38h06a4308_0
python=3.8.16=h7a1cb2a_3
python-dateutil=2.8.2=pyhd3eb1b0_0
python_abi=3.8=2_cp38
pytorch=2.0.0=py3.8_cuda11.7_cudnn8.5.0_0
pytorch-cuda=11.7=h778d358_3
pytorch-mutex=1.0=cuda
pytz=2022.7=py38h06a4308_0
pyzmq=19.0.2=py38ha71036d_2
qt-main=5.15.2=h8373d8f_8
qt-webengine=5.15.9=hbbf29b9_6
qtwebkit=5.212=h3fafdc1_5
re2=2022.04.01=h295c915_0
readline=8.2=h5eee18b_0
requests=2.29.0=py38h06a4308_0
requests-oauthlib=1.3.0=py_0
rsa=4.7.2=pyhd3eb1b0_1
scikit-learn=1.2.2=py38h6a678d5_0
scipy=1.10.1=py38h14f4228_0
seaborn=0.12.2=py38h06a4308_0
setuptools=66.0.0=py38h06a4308_0
sip=6.6.2=py38h6a678d5_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.41.2=h5eee18b_0
stack_data=0.6.2=pyhd8ed1ab_0
sympy=1.11.1=py38h06a4308_0
tensorboard=2.11.0=py38h06a4308_0
tensorboard-data-server=0.6.1=py38h52d8a92_0
tensorboard-plugin-wit=1.8.1=py38h06a4308_0
threadpoolctl=2.2.0=pyh0d69192_0
tk=8.6.12=h1ccaba5_0
toml=0.10.2=pyhd3eb1b0_0
torchaudio=2.0.0=py38_cu117
torchtriton=2.0.0=py38
torchvision=0.15.0=py38_cu117
tornado=6.1=py38h0a891b7_3
tqdm=4.65.0=py38hb070fc8_0
traitlets=5.9.0=pyhd8ed1ab_0
typing_extensions=4.5.0=py38h06a4308_0
urllib3=1.26.15=py38h06a4308_0
wcwidth=0.2.6=pyhd8ed1ab_0
werkzeug=2.2.3=py38h06a4308_0
wheel=0.38.4=py38h06a4308_0
xz=5.2.10=h5eee18b_1
yarl=1.8.1=py38h5eee18b_0
zeromq=4.3.4=h9c3ff4c_1
zipp=3.11.0=py38h06a4308_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.5=hc292b87_0
